v1

+++ src/main/java/com/vstu/entity/Certification.java
@@ -1,7 +1,12 @@
 package com.vstu.entity;
 
 import java.io.Serializable;
-import javax.persistence.*;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
 
 
 /**
@@ -14,7 +19,7 @@ public class Certification implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	private int id;
+	private long id;
 
 	//bi-directional many-to-one association to Plan
 	@ManyToOne
@@ -24,11 +29,11 @@ public class Certification implements Serializable {
 	public Certification() {
 	}
 
-	public int getId() {
+	public long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
