+++ src/main/java/com/vstu/entity/WeeksSemestr.java
@@ -1,7 +1,14 @@
 package com.vstu.entity;
 
 import java.io.Serializable;
-import javax.persistence.*;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.Table;
 
 
 /**
@@ -15,7 +22,7 @@ public class WeeksSemestr implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	private int id;
+	private long id;
 
 	@Column(name="count_weeks")
 	private int countWeeks;
@@ -31,11 +38,11 @@ public class WeeksSemestr implements Serializable {
 	public WeeksSemestr() {
 	}
 
-	public int getId() {
+	public long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
